# FrontEnd/ChamberAppel/src/app/apis/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Chamber d'appel",
    "description": "Ceci est une documentation d'API Swagger pour la gestion de la Chambre d'Appel de la Cour des Comptes.",
    "contact": {
      "name": "Cour des comptes",
      "url": "https://www.courdescomptes.ma"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Permissions/GetAllFiltredAsync": {
      "post": {
        "tags": [
          "Permissions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreMotsCleDatatableRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreMotsCleDatatableRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreMotsCleDatatableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDatatableResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDatatableResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDatatableResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Permissions/GetByIdAsync/{id}": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Permissions/CreateAsync": {
      "post": {
        "tags": [
          "Permissions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Permissions/UpdateAsync": {
      "put": {
        "tags": [
          "Permissions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Permissions/DeleteAsync/{id}": {
      "delete": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Permissions/ExporterAsync": {
      "post": {
        "tags": [
          "Permissions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreMotsCle"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreMotsCle"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreMotsCle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BooleanApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "boolean"
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FluentValidationErrors"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DtoFiltreMotsCle": {
        "type": "object",
        "properties": {
          "motsCle": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DtoFiltreMotsCleDatatableRequest": {
        "type": "object",
        "properties": {
          "filtre": {
            "$ref": "#/components/schemas/DtoFiltreMotsCle"
          },
          "pagination": {
            "$ref": "#/components/schemas/DtoPagination"
          }
        },
        "additionalProperties": false
      },
...
      "DtoPagination": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderByDirection": {
            "type": "string",
            "nullable": true
          },
          "orderBy": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FluentValidationErrors": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
      "Permission": {
        "required": [
          "groupe",
          "label"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "label": {
            "maxLength": 150,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "integer",
            "format": "int32"
          },
          "groupe": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PermissionApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Permission"
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FluentValidationErrors"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PermissionDatatableResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "nullable": true
          },
          "recordTotal": {
            "type": "integer",
            "format": "int32"
          },
          "recordFiltred": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}